Update to 3.2.1

+++ source/entity/mobs/nether/EntityFakePigman.java
@@ -92,7 +92,7 @@ public class EntityFakePigman extends AoAMeleeMob {
     }
 
     @Override
-    protected float getSpawnChanceFactor() {
+    protected double getSpawnChanceFactor() {
         return 0.5f;
     }
 }
