+++ source/entity/mobs/nether/EntityHellspot.java
@@ -93,7 +93,7 @@ public class EntityHellspot extends AoAMeleeMob implements SpecialPropertyEntity
     }
 
     @Override
-    protected float getSpawnChanceFactor() {
+    protected double getSpawnChanceFactor() {
         return 0.5f;
     }
