+++ source/entity/mobs/nether/EntityInfernal.java
@@ -115,7 +115,7 @@ public class EntityInfernal extends AoAMeleeMob implements SpecialPropertyEntity
     }
 
     @Override
-    protected float getSpawnChanceFactor() {
+    protected double getSpawnChanceFactor() {
         return 0.5f;
     }
